refactor: remove message-content-repository, use message-repository only
