+++ apps/etl-func/src/adapters/message.ts
@@ -1,13 +1,13 @@
-import {
-  GetMessageByMetadataReturnType,
-  MessageMetadata,
-  MessageRepository,
-} from "@/domain/entities/message.js";
 import { RestError } from "@azure/storage-blob";
 import { Logger } from "pino";
 import * as z from "zod";
 
 import { BlobMessageContent } from "./blob-storage/message-content.js";
+import {
+  GetMessageByMetadataReturnType,
+  MessageMetadata,
+  MessageRepository,
+} from "@/domain/message.js";
 
 export class MessageAdapter implements MessageRepository {
   #content: BlobMessageContent;
